# paperless-backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Paperless REST API
  description: |
    Document Management System REST API for the SWEN3 Paperless project.
    Provides endpoints for:
    - Document Upload: file storage, automated OCR processing, Elasticsearch full-text indexing, and GenAI summarization.
    - Document Search: full-text and fuzzy search powered by Elasticsearch.
    - Document Management: document lifecycle (details, metadata update, deletion, download) and metadata (document types).
    - User and Team Management: authentication (login, me), user management (User-MM), teams, and role-based memberships (User <-- m:n (role) --> Team; roles: ADMIN, READONLY, READ_WRITE).
  version: 1.0.0
  x-api-id: d0184f38-b98d-11e7-9c56-68f728c1ba70
  x-audience: company-internal
  contact:
    name: Philip Duong & Patrick Hornek
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local Docker / Development Server
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List all users
      description: Returns a list of all registered users in the application.
      operationId: getUsers
      responses:
        '200':
          description: List of users retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Public endpoint to create/register a new user account (User-MM). Usernames must be unique; registration with an existing username returns 409.
      operationId: createUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully.
          headers:
            Location:
              description: URI of the newly created user resource.
              schema:
                type: string
                format: uri
                example: "/users/550e8400-e29b-41d4-a716-446655440000"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Returns user profile by ID.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: User retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Users
      summary: Update user
      description: Updates only the authenticated user's own profile; updating another user's account is forbidden. A supplied username must be unique across accounts; an already-used username returns 409 Conflict.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes only the authenticated user's own account; deleting another user's account is forbidden.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '204':
          description: User deleted successfully.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{id}/teams:
    get:
      tags:
        - Users
      summary: Get teams and roles for user
      description: Returns the authenticated user's teams and roles; requesting memberships for another user is forbidden.
      operationId: getUserTeams
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: User team memberships and roles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserTeamMembershipListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # TEAMS & ROLES (User <-- m:n (role) --> Team)
  # ============================================================================
  /teams:
    get:
      tags:
        - Teams
      summary: List all teams
      description: Returns only teams the authenticated user belongs to; the list is empty when the user has no team memberships.
      operationId: getTeams
      responses:
        '200':
          description: List of teams retrieved successfully; may be empty when the user has no memberships.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Teams
      summary: Create a new team
      description: Creates a new team; the authenticated creator becomes its immutable owner.
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Team created successfully.
          headers:
            Location:
              description: URI of the newly created team resource.
              schema:
                type: string
                format: uri
                example: "/teams/550e8400-e29b-41d4-a716-446655440001"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Standard Authorization Bearer header using an opaque access token.
  parameters:
    PathId:
      name: id
      in: path
      description: Unique identifier of the resource.
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Role:
      type: string
      description: |-
        READONLY can read, search, view, and download team documents.
        READ_WRITE can also upload, edit, and delete team documents.
        ADMIN has READ_WRITE permissions and additionally manages team memberships; the owner cannot be removed.
      enum:
        - ADMIN
        - READONLY
        - READ_WRITE
      example: ADMIN

    # --- User Schemas (User-MM & Login) ---
    UserResponse:
      type: object
      required:
        - id
        - username
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        username:
          type: string
          description: Unique username that the account owner may change to another unused username.
          example: john_doe
        created_at:
          type: string
          format: date-time
          example: "2026-10-15T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-10-15T14:32:10Z"
    UserListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Required unique username of at least 4 characters; usernames may later be changed by the account owner to another unused username.
          minLength: 4
          maxLength: 50
          example: john_doe
        password:
          type: string
          description: Raw password of at least 8 characters.
          format: password
          minLength: 8
          maxLength: 100
          example: "SecretP@ss123"
    UpdateUserRequest:
      type: object
      properties:
        username:
          type: string
          description: Optional new username of at least 4 characters; it must be unique across accounts or the update returns 409 Conflict.
          minLength: 4
          maxLength: 50
          example: john_doe
        password:
          type: string
          description: Optional new raw password of at least 8 characters; omit to leave the current password unchanged.
          format: password
          minLength: 8
          maxLength: 100
          example: "NewSecretP@ss123"
...
    TeamResponse:
      type: object
      required:
        - id
        - name
        - owner_id
        - created_at
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        name:
          type: string
          example: Accounting
        owner_id:
          type: string
          format: uuid
          description: User ID of the team creator and immutable owner.
          example: "550e8400-e29b-41d4-a716-446655440000"
        description:
          type: string
          nullable: true
          example: Handles invoices and bills
        created_at:
          type: string
          format: date-time
          example: "2026-10-15T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-10-15T14:32:10Z"
    TeamListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TeamResponse'
    CreateTeamRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Required nonblank team name.
          minLength: 1
          maxLength: 100
          example: Accounting
        description:
          type: string
          nullable: true
          maxLength: 255
          example: Handles invoices and bills
...
    UserTeamMembershipResponse:
      type: object
      required:
        - team
        - role
      properties:
        team:
          $ref: '#/components/schemas/TeamResponse'
        role:
          $ref: '#/components/schemas/Role'
    UserTeamMembershipListResponse:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserTeamMembershipResponse'

    # --- Document Schemas ---
...
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: "urn:problem-type:not-found"
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: "Resource Not Found"
        status:
          type: integer
          description: The HTTP status code generated by the origin server.
          example: 404
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: "Document with id 550e8400-e29b-41d4-a716-446655440002 was not found."
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
          example: "/documents/550e8400-e29b-41d4-a716-446655440002"
        invalid_params:
          type: array
          description: An array of invalid parameter errors (RFC 7807 extension).
          items:
            $ref: '#/components/schemas/InvalidParam'
    InvalidParam:
      type: object
      required:
        - name
        - reason
      properties:
        name:
          type: string
          description: The name of the invalid parameter or field.
          example: "title"
        reason:
          type: string
          description: Explanation of why the parameter is invalid.
          example: "Title must not be blank."

  # ============================================================================
  # RESPONSES
  # ============================================================================
  responses:
    BadRequestError:
      description: Bad Request - Validation failure or malformed payload.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnauthorizedError:
      description: Unauthorized - Authentication is required or credentials are invalid.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    ForbiddenError:
      description: Forbidden - The authenticated user is not authorized to perform this operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFoundError:
      description: Not Found - The requested entity does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    ConflictError:
      description: Conflict - A resource with the same identifier or unique constraint already exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    InternalServerError:
      description: Internal Server Error - Unexpected failure on the server.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'